+++ CHANGELOG.md
@@ -132,6 +132,7 @@ This view now also sorts events in reverse chronological order.
 - Indexing jobs will now output a nicer error message if things go “expectedly” wrong ([\#431](https://gitlab.desy.de/cfel-sc/amarcord-parent/amarcord-serial/-/merge_requests/431))
 - Indexing jobs now show the resulting `.stream` file ([\#444](https://gitlab.desy.de/cfel-sc/amarcord-parent/amarcord-serial/-/issues/444))
 - API: When creating (or updating) a run, you can instruct it to create a Data Set for the run as well ([#457](https://gitlab.desy.de/cfel-sc/amarcord-parent/amarcord-serial/-/issues/457))
+- API: You can now create a finished indexing result ([#458](https://gitlab.desy.de/cfel-sc/amarcord-parent/amarcord-serial/-/issues/458))
 
 ## Fixes